requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
http-ece==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
py-vapid==1.8.2
pycparser==2.21
python-dotenv==0.21.0
pywebpush==1.14.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
